Создал MovieQuizPresenter и перенес туда метод convert и все необходимые свойства. Исправил код и проверил работоспособность программы
