+++ MovieQuiz/Presentation/MovieQuizPresenter.swift
@@ -1,0 +1,33 @@
+import UIKit
+
+final class MovieQuizPresenter {
+    
+    
+    
+    let questionAmount: Int = 10
+    private var currentQuestionIndex = 0
+    
+    func isLastQuestion() -> Bool {
+        currentQuestionIndex == questionAmount - 1
+    }
+
+    func resetQuestionIndex() {
+        currentQuestionIndex = 0
+    }
+    
+    func switchToNextQuestion() {
+        currentQuestionIndex += 1
+    }
+    
+    // функция конвертации
+    
+    func convert(model: QuizQuestion) -> QuizStepViewModel {
+        
+        let questionStep = QuizStepViewModel(image: UIImage(data: model.image) ?? UIImage(),
+                                             question: model.text,
+                                             questionNumber: "\(currentQuestionIndex + 1)/\(questionAmount)")
+        return questionStep
+        
+    }
+    
+}
